Workaround build issues.
- 'Complex' causes all sorts of issues, I've replaced it with
  std::complex<float> in the past, but just to get it in a working state
  I've swapped it out with Complex_.
- Fix a non-lvalue increment expression
- Fix a doubly qualified constructor (I'm not sure how this compiled
  when it was written)

+++ src/alsaio.cxx
@@ -421,7 +421,7 @@ void help(const char* appl)
 "-e,--effect    apply an effect (bandpass filter sweep)\n"
 ,appl);
         printf("Recognized sample formats are:");
-        for (k = 0; k < SND_PCM_FORMAT_LAST; ++(unsigned long) k) {
+        for (k = 0; k < SND_PCM_FORMAT_LAST; ++k) {
                 const char *s = snd_pcm_format_name((snd_pcm_format_t) k);
                 if (s)
                         printf(" %s", s);
